+++ X10D.Unity.Tests/Assets/Tests/Numerics/Vector4Tests.cs
@@ -2,12 +2,56 @@
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
+using X10D.Core;
 using X10D.Unity.Numerics;
+using Random = System.Random;
 
 namespace X10D.Unity.Tests.Numerics
 {
     public class Vector4Tests
     {
+        [UnityTest]
+        public IEnumerator ToSystemVector_ShouldReturnVector_WithEqualComponents()
+        {
+            var random = new Random();
+            float x = random.NextSingle();
+            float y = random.NextSingle();
+            float z = random.NextSingle();
+            float w = random.NextSingle();
+
+            var vector = new Vector4(x, y, z, w);
+            var systemVector = vector.ToSystemVector();
+
+            Assert.AreEqual(vector.magnitude, systemVector.Length(), 1e-6f);
+            Assert.AreEqual(vector.x, systemVector.X, 1e-6f);
+            Assert.AreEqual(vector.y, systemVector.Y, 1e-6f);
+            Assert.AreEqual(vector.z, systemVector.Z, 1e-6f);
+            Assert.AreEqual(vector.w, systemVector.W, 1e-6f);
+
+            yield break;
+        }
+
+        [UnityTest]
+        public IEnumerator ToUnityVector_ShouldReturnVector_WithEqualComponents()
+        {
+            var random = new Random();
+            float x = random.NextSingle();
+            float y = random.NextSingle();
+            float z = random.NextSingle();
+            float w = random.NextSingle();
+
+            var vector = new System.Numerics.Vector4(x, y, z, w);
+            var unityVector = vector.ToUnityVector();
+
+            Assert.AreEqual(vector.Length(), unityVector.magnitude, 1e-6f);
+            Assert.AreEqual(vector.X, unityVector.x, 1e-6f);
+            Assert.AreEqual(vector.Y, unityVector.y, 1e-6f);
+            Assert.AreEqual(vector.Z, unityVector.z, 1e-6f);
+            Assert.AreEqual(vector.W, unityVector.w, 1e-6f);
+
+            yield break;
+        }
+
         [UnityTest]
         public IEnumerator WithW_ShouldReturnVectorWithNewW_GivenVector()
         {
@@ -24,7 +68,7 @@ namespace X10D.Unity.Tests.Numerics
             Assert.AreEqual(new Vector4(1, 0, 0, 1), new Vector4(1, 0, 0, 0).WithW(1));
             Assert.AreEqual(new Vector4(0, 1, 0, 1), new Vector4(0, 1, 0, 0).WithW(1));
             Assert.AreEqual(new Vector4(0, 0, 1, 1), new Vector4(0, 0, 1, 0).WithW(1));
-            
+
             yield break;
         }
 
@@ -44,7 +88,7 @@ namespace X10D.Unity.Tests.Numerics
             Assert.AreEqual(new Vector4(1, 0, 0, 0), new Vector4(1, 0, 0, 0).WithX(1));
             Assert.AreEqual(new Vector4(1, 1, 0, 0), new Vector4(0, 1, 0, 0).WithX(1));
             Assert.AreEqual(new Vector4(1, 0, 1, 0), new Vector4(0, 0, 1, 0).WithX(1));
-            
+
             yield break;
         }
 
@@ -64,7 +108,7 @@ namespace X10D.Unity.Tests.Numerics
             Assert.AreEqual(new Vector4(1, 1, 0, 0), new Vector4(1, 0, 0, 0).WithY(1));
             Assert.AreEqual(new Vector4(0, 1, 0, 0), new Vector4(0, 1, 0, 0).WithY(1));
             Assert.AreEqual(new Vector4(0, 1, 1, 0), new Vector4(0, 0, 1, 0).WithY(1));
-            
+
             yield break;
         }
 
@@ -84,7 +128,7 @@ namespace X10D.Unity.Tests.Numerics
             Assert.AreEqual(new Vector4(1, 0, 1, 0), new Vector4(1, 0, 0, 0).WithZ(1));
             Assert.AreEqual(new Vector4(0, 1, 1, 0), new Vector4(0, 1, 0, 0).WithZ(1));
             Assert.AreEqual(new Vector4(0, 0, 1, 0), new Vector4(0, 0, 1, 0).WithZ(1));
-            
+
             yield break;
         }
     }