+++ X10D.Unity/src/Numerics/Vector3Extensions.cs
@@ -10,6 +10,30 @@ namespace X10D.Unity.Numerics;
 public static class Vector3Extensions
 {
     /// <summary>
+    ///     Converts the current vector to a <see cref="System.Numerics.Vector3" />.
+    /// </summary>
+    /// <param name="vector">The vector to convert.</param>
+    /// <returns>The converted vector.</returns>
+    [Pure]
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static System.Numerics.Vector3 ToSystemVector(this Vector3 vector)
+    {
+        return new System.Numerics.Vector3(vector.x, vector.y, vector.z);
+    }
+
+    /// <summary>
+    ///     Converts the current vector to a <see cref="Vector3" />.
+    /// </summary>
+    /// <param name="vector">The vector to convert.</param>
+    /// <returns>The converted vector.</returns>
+    [Pure]
+    [MethodImpl(MethodImplOptions.AggressiveInlining)]
+    public static Vector3 ToUnityVector(this System.Numerics.Vector3 vector)
+    {
+        return new Vector3(vector.X, vector.Y, vector.Z);
+    }
+
+    /// <summary>
     ///     Returns a vector whose Y and Z components are the same as the specified vector, and whose X component is a new value.
     /// </summary>
     /// <param name="vector">The vector to copy.</param>